webapp/requirements.txt
coreapi==2.3.3
coreschema==0.0.4
Django==2.0.4
django-allauth==0.35.0
django-cors-headers==2.2.0
django-filter==1.1.0
django-rest-auth==0.9.3
djangorestframework==3.8.2
feedparser==5.2.1
gunicorn==19.7.1
gevent==1.2.1
Markdown==2.6.11
psycopg2-binary==2.7.4
Pygments==2.2.0